requirements.txt
Django>=4.2.6
django-ckeditor>=6.7.0
Pillow>=10.0.0
django-image-cropping>=1.5.0
easy-thumbnails>=2.8.0

